code_0917_zhi/requirements.txt
flask
flask-mail
tinydb
bcrypt
python-dateutil
python-dotenv
openai

# Optional dependencies (install only if using speech recognition)
# SpeechRecognition
# PyAudio
bcrypt
flask
flask-mail
openai
python-dateutil
python-dotenv
tinydb
